Fix errors in -std=c++11 mode
`free_operators.hpp` had a couple of instances of narrowing conversions
within braced initializers, which is ill-formed since C++11.

`type_traits.hpp` had two instances of aggregates being
value-initialized using empty parentheses. Since these aggregates had
implicitly deleted default constructors (due to the presence of `const`
data members), value-initialization performs default-initialization,
which is ill-formed. We need aggregate-initialization to be performed,
so the initialization must use braces instead of parentheses for C++11
and later.


Cherry-picked from the develop branch

+++ unittests/data/free_operators.hpp
@@ -32,12 +32,12 @@ bool operator!( const number& x ){
 }
 
 number operator*( const number& n,  double i ){
-    number n2 = { n.i * i };
+    number n2 = { static_cast<int>(n.i * i) };
     return n2;
 }
 
 number operator*( double i, const number& n ){
-    number n2 = { n.i * i };
+    number n2 = { static_cast<int>(n.i * i) };
     return n2;
 }
